+++ actiontext/test/dummy/app/models/message.rb
@@ -1,4 +1,6 @@
 class Message < ApplicationRecord
+  include ActionText::Attachable
+
   has_rich_text :content
   has_rich_text :body