+++ actiontext/CHANGELOG.md
@@ -1,3 +1,16 @@
+*   Add `action_text_attachment` helper to `ActiveRecord::FixtureSet` context to
+    generate an `<action-text-attachment sgid="..."></action-text-attachment>`
+    element with a valid `sgid` attribute.
+
+    ```ruby
+    hello_world_review_content:
+      record: hello_world (Review)
+      name: content
+      body: <p><%= action_text_attachment("messages", :hello_world) %> is great!</p>
+    ```
+
+    *Sean Doyle*
+
 *   Locate `fill_in_rich_text_area` by `<label>` text
 
     In addition to searching for `<trix-editor>` elements with the appropriate